# backend/ms-geolocalisation/src/main/resources/openapi/api.yml
openapi: 3.0.3
info:
  title: PGVPT - MS Géolocalisation API
  description: Propriétaire des coordonnées, du référentiel administratif, des zones géographiques et touristiques, des points
    d'accès et des recherches spatiales. Une géolocalisation cible exactement un patrimoine ou une entreprise; les références
    sont des UUID.
  version: 14.0.0
  contact:
    name: PGVPT
    email: contact@pgvpt.sn
servers:
  - url: http://localhost:3002/api/v1/geolocalisations
    description: Environnement local
  - url: https://api.pgvpt.sn/geolocalisation/api/v1
    description: Environnement de production
...
paths:
  /geolocalisations:
    get:
      tags:
        - Geolocalisations
      summary: Lister les Geolocalisations
      operationId: getGeolocalisations
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Size'
        - name: departement
          in: query
          schema:
            type: string
        - name: commune
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Liste des Geolocalisations
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageGeolocalisation'
        '500':
          $ref: '#/components/responses/InternalServerError'
    post:
      tags:
        - Geolocalisations
      summary: Créer une géolocalisation
      operationId: createGeolocalisation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GeolocalisationCreate'
      responses:
        '201':
          description: Géolocalisation créée
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Geolocalisation'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalServerError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
      parameters: []
      security:
        - bearerAuth: []
  /geolocalisations/{id}:
    get:
      tags:
        - Geolocalisations
      summary: Obtenir une géolocalisation
      operationId: getGeolocalisation
      responses:
        '200':
          description: Géolocalisation trouvée
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Geolocalisation'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    put:
      tags:
        - Geolocalisations
      summary: Modifier une géolocalisation
      operationId: updateGeolocalisation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GeolocalisationUpdate'
      responses:
        '200':
          description: Géolocalisation modifiée
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Geolocalisation'
        '500':
          $ref: '#/components/responses/InternalServerError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
      security:
        - bearerAuth: []
    patch:
      tags:
        - Geolocalisations
      summary: Mise à jour partielle d'une géolocalisation
      operationId: patchGeolocalisation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GeolocalisationUpdate'
      responses:
        '200':
          description: Géolocalisation modifiée
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Geolocalisation'
        '500':
          $ref: '#/components/responses/InternalServerError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
      security:
        - bearerAuth: []
    delete:
      tags:
        - Geolocalisations
      summary: Supprimer une géolocalisation
      operationId: deleteGeolocalisation
      responses:
        '204':
          description: Géolocalisation supprimée
        '500':
          $ref: '#/components/responses/InternalServerError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
      security:
        - bearerAuth: []
    parameters:
      - $ref: '#/components/parameters/GeolocalisationId'
  /patrimoines/{patrimoineId}/geolocalisation:
    get:
      tags:
        - Geolocalisations
      summary: Obtenir la géolocalisation d'un patrimoine
      operationId: getPatrimoineGeolocalisation
      responses:
        '200':
          description: Géolocalisation du patrimoine
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Geolocalisation'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    put:
      tags:
        - Geolocalisations
      summary: Mettre à jour la géolocalisation d'un patrimoine
      operationId: updatePatrimoineGeolocalisation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GeolocalisationUpdate'
      responses:
        '200':
          description: Géolocalisation mise à jour
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Geolocalisation'
        '500':
          $ref: '#/components/responses/InternalServerError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
      security:
        - bearerAuth: []
    patch:
      tags:
        - Geolocalisations
      summary: Mise à jour partielle de la géolocalisation d'un patrimoine
      operationId: patchPatrimoineGeolocalisation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GeolocalisationUpdate'
      responses:
        '200':
          description: Géolocalisation mise à jour
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Geolocalisation'
        '500':
          $ref: '#/components/responses/InternalServerError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
      security:
        - bearerAuth: []
    parameters:
      - $ref: '#/components/parameters/PatrimoineId'
...
components:
  parameters:
...
    GeolocalisationId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    PatrimoineId:
      name: patrimoineId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    Page:
      name: page
      in: query
      required: false
      description: Numéro de page, indexé à partir de 0.
      schema:
        type: integer
        minimum: 0
        default: 0
    Size:
      name: size
      in: query
      required: false
      description: Nombre maximal d’éléments par page.
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
  responses:
    BadRequest:
      description: Requête invalide
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    NotFound:
      description: Ressource introuvable
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    Conflict:
      description: Conflit métier
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    Unauthorized:
      description: Authentification requise
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    Forbidden:
      description: Accès interdit
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    InternalServerError:
      description: Erreur interne
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
...
  schemas:
    BaseEntity:
      type: object
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        createdAt:
          type: string
          format: date-time
          readOnly: true
        updatedAt:
          type: string
          format: date-time
          readOnly: true
      description: Classe racine commune de toutes les entités persistantes du PGVPT.
    Geolocalisation:
      allOf:
        - $ref: '#/components/schemas/BaseEntity'
        - type: object
          required:
            - id
            - zoneTouristiqueId
            - latitude
            - longitude
          properties:
            latitude:
              type: number
              format: double
              minimum: -90
              maximum: 90
            longitude:
              type: number
              format: double
              minimum: -180
              maximum: 180
            altitude:
              type: number
              format: double
            precisionMetres:
              type: number
              format: double
              minimum: 0
            adresse:
              type: string
            lieuDit:
              type: string
            repere:
              type: string
            systemeReference:
              type: string
              default: WGS84
              example: WGS84
            codeEpsg:
              type: integer
              default: 4326
              example: 4326
            source:
              type: string
              enum:
                - GPS
                - ADMINISTRATION
                - SIG
                - OPEN_DATA
                - IMPORT
                - AUTRE
              description: Origine de la donnée de géolocalisation.
            dateAcquisition:
              type: string
              format: date
            methodeAcquisition:
              type: string
              enum:
                - GPS
                - RELEVE_TERRAIN
                - SATELLITE
                - DRONE
                - GEOCODAGE
                - IMPORT_SIG
                - AUTRE
            niveauFiabilite:
              type: string
              enum:
                - FAIBLE
                - MOYEN
                - ELEVE
                - TRES_ELEVE
            zoneTouristiqueId:
              type: string
              format: uuid
            villageId:
              type: string
              format: uuid
              nullable: true
            quartierId:
              type: string
              format: uuid
              nullable: true
            patrimoineId:
              type: string
              format: uuid
              nullable: true
              description: Identifiant du patrimoine provenant de ms-patrimoine.
            entrepriseId:
              type: string
              format: uuid
              nullable: true
              description: Identifiant de l'entreprise provenant de ms-entreprise.
          oneOf:
            - required:
                - patrimoineId
              not:
                required:
                  - entrepriseId
            - required:
                - entrepriseId
              not:
                required:
                  - patrimoineId
          description: 'Entité persistante décrivant la localisation d''une ressource géolocalisable. Elle hérite de BaseEntity.
          Une géolocalisation est associée à exactement une ressource : soit un patrimoine provenant de ms-patrimoine, soit
          une entreprise provenant de ms-entreprise.'
    GeolocalisationCreate:
      type: object
      required:
        - zoneTouristiqueId
        - latitude
        - longitude
      properties:
        latitude:
          type: number
          format: double
          minimum: -90
          maximum: 90
        longitude:
          type: number
          format: double
          minimum: -180
          maximum: 180
        altitude:
          type: number
          format: double
        precisionMetres:
          type: number
          format: double
          minimum: 0
        adresse:
          type: string
        lieuDit:
          type: string
        repere:
          type: string
        systemeReference:
          type: string
          default: WGS84
          example: WGS84
        codeEpsg:
          type: integer
          default: 4326
          example: 4326
        source:
          type: string
          enum:
            - GPS
            - ADMINISTRATION
            - SIG
            - OPEN_DATA
            - IMPORT
            - AUTRE
          description: Origine de la donnée de géolocalisation.
        dateAcquisition:
          type: string
          format: date
        methodeAcquisition:
          type: string
          enum:
            - GPS
            - RELEVE_TERRAIN
            - SATELLITE
            - DRONE
            - GEOCODAGE
            - IMPORT_SIG
            - AUTRE
        niveauFiabilite:
          type: string
          enum:
            - FAIBLE
            - MOYEN
            - ELEVE
            - TRES_ELEVE
        zoneTouristiqueId:
          type: string
          format: uuid
        villageId:
          type: string
          format: uuid
          nullable: true
        quartierId:
          type: string
          format: uuid
          nullable: true
        patrimoineId:
          type: string
          format: uuid
          nullable: true
          description: Identifiant du patrimoine provenant de ms-patrimoine.
        entrepriseId:
          type: string
          format: uuid
          nullable: true
          description: Identifiant de l'entreprise provenant de ms-entreprise.
      oneOf:
        - required:
            - patrimoineId
          not:
            required:
              - entrepriseId
        - required:
            - entrepriseId
          not:
            required:
              - patrimoineId
      description: 'Une géolocalisation doit référencer exactement une ressource : un patrimoine ou une entreprise.'
    GeolocalisationUpdate:
      type: object
      properties:
        latitude:
          type: number
          format: double
          minimum: -90
          maximum: 90
        longitude:
          type: number
          format: double
          minimum: -180
          maximum: 180
        altitude:
          type: number
          format: double
        precisionMetres:
          type: number
          format: double
          minimum: 0
        adresse:
          type: string
        lieuDit:
          type: string
        repere:
          type: string
        systemeReference:
          type: string
          default: WGS84
          example: WGS84
        codeEpsg:
          type: integer
          default: 4326
          example: 4326
        source:
          type: string
          enum:
            - GPS
            - ADMINISTRATION
            - SIG
            - OPEN_DATA
            - IMPORT
            - AUTRE
          description: Origine de la donnée de géolocalisation.
        dateAcquisition:
          type: string
          format: date
        methodeAcquisition:
          type: string
          enum:
            - GPS
            - RELEVE_TERRAIN
            - SATELLITE
            - DRONE
            - GEOCODAGE
            - IMPORT_SIG
            - AUTRE
        niveauFiabilite:
          type: string
          enum:
            - FAIBLE
            - MOYEN
            - ELEVE
            - TRES_ELEVE
        zoneTouristiqueId:
          type: string
          format: uuid
        villageId:
          type: string
          format: uuid
          nullable: true
        quartierId:
          type: string
          format: uuid
          nullable: true
        patrimoineId:
          type: string
          format: uuid
          nullable: true
          description: Identifiant du patrimoine provenant de ms-patrimoine.
        entrepriseId:
          type: string
          format: uuid
          nullable: true
          description: Identifiant de l'entreprise provenant de ms-entreprise.
      oneOf:
        - required:
            - patrimoineId
          not:
            required:
              - entrepriseId
        - required:
            - entrepriseId
          not:
            required:
              - patrimoineId
      description: 'Une géolocalisation doit référencer exactement une ressource : un patrimoine ou une entreprise.'
...
    PageGeolocalisation:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/Geolocalisation'
        page:
          type: integer
          minimum: 0
        size:
          type: integer
          minimum: 1
        totalElements:
          type: integer
          format: int64
          minimum: 0
        totalPages:
          type: integer
          minimum: 0
        first:
          type: boolean
        last:
          type: boolean
      required:
        - content
        - page
        - size
        - totalElements
        - totalPages
        - first
        - last
      description: Résultat paginé des géolocalisations.
    ApiError:
      type: object
      required:
        - timestamp
        - status
        - code
        - message
        - path
      properties:
        timestamp:
          type: string
          format: date-time
        status:
          type: integer
        code:
          type: string
        message:
          type: string
        path:
          type: string
        details:
          type: object
          additionalProperties: true
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []